pulled registration actions out of user and static pages controller
