+++ app/controllers/static_pages_controller.rb
@@ -3,10 +3,10 @@ class StaticPagesController < ApplicationController
                      only: [:api, :copyright, :home, :status]
 
   skip_before_filter :registration,
-                     only: [:api, :copyright, :home, :status, :verification_sent]
+                     only: [:api, :copyright, :home, :status]
 
   fine_print_skip :general_terms_of_use, :privacy_policy,
-                  only: [:api, :copyright, :home, :status, :verification_sent]
+                  only: [:api, :copyright, :home, :status]
 
   skip_protect_beta :only => [:status]
 
@@ -27,9 +27,6 @@ class StaticPagesController < ApplicationController
       store_url # needed for happy login flow, authenticate_user! does it too
       redirect_to login_path
     end
-  end
-
-  def verification_sent
   end
 
   # Used by AWS (and others) to make sure the site is still up.
