+++ app/views/contact_infos/confirm_unclaimed.html.erb
@@ -9,7 +9,7 @@
   <% if @handler_result.outputs.can_log_in %>
     You can now <%= link_to("sign in", login_url) %> to your new account.
   <% else %>
-    Please <%= link_to('sign up',register_url) %> for a account using email <%= @handler_result.outputs.email_address %> address in order to complete the process.
+    Please <%= link_to('sign up',registration_complete_url) %> for a account using email <%= @handler_result.outputs.email_address %> address in order to complete the process.
   <% end %>
 </p>
 <% end %>
