+++ app/views/registration/complete.html.erb
@@ -2,7 +2,7 @@
 
 <p>* indicates required field.</p>
 
-<%= lev_form_for :register, url: register_path, method: :put, html: {id: 'profile-form'} do |f| %>
+<%= lev_form_for :register, url: registration_complete_path, method: :put, html: {id: 'profile-form'} do |f| %>
 
   <%= standard_field form: f, name: :username, type: :text_field, label: "Username *",
                      options: {value: current_user.username} %>
