+++ app/views/sessions/ask_new_or_returning.html.erb
@@ -28,7 +28,7 @@ profile.</p>
     </div>
     <div class='well'>
     <%= link_to "Continue".html_safe,
-                register_path, class: 'standard' %>
+                registration_complete_path, class: 'standard' %>
     </div>
 
   </div>
