+++ config/routes.rb
@@ -27,11 +27,15 @@ Accounts::Application.routes.draw do
   scope controller: 'users' do
     get 'profile', action: :edit
     put 'profile', action: :update
-    get 'register'
-    put 'register'
     get 'ask_for_email'
     put 'ask_for_email'
-    put 'i_clicked_verification_link'
+  end
+
+  namespace 'registration' do
+    get 'complete'
+    put 'complete'
+    get 'verification_pending'
+    put 'i_verified'
   end
 
   resource :identity, only: :update
@@ -64,7 +68,6 @@ Accounts::Application.routes.draw do
   scope controller: 'static_pages' do
     get 'copyright'
     get 'status'
-    get 'verification_sent'
   end
 
   apipie
